front: operationalstudies: focus added on the name input when modal opens
- added focus on mapSearchStation
- added focus on addOrEditProjectModal
- added focus on addOrEditStudyModal
- added focus on addOrEditScenarioModal
- create custom hook useModalFocusTrap in utils/hooks
